Split similarity background compute work
